Add download feature
Images as well as files with most common extensions can be saved
directly from app.

If the access to read and write on external storage is granted, the file
is saved in the `/storage/emulated/0/Download` folder. If denied (or if
the folder doesn't exist), it is saved in the app `files` folder.