Use ChangeNotifier for Settings
Since settings are loaded asynchronously, they require a controller. The
drawback is that only the `TextEditingController` is available. In case
we need to introduce another widget type (e.g. a dropdown list), we are
stuck.

To circumvent the limitation, we use a ChangeNotifier which allows
asynchronous loading while giving the possibility to use different
widget types.