Add button to fetch original content
Miniflux provides us 2 APIs: one to fetch the original content, and one
to update an entity. We update the content using these APIs. We first
download the original content from Miniflux and, if that succeeds, we
PUT an update to that entity and replace the local content.

Due to entity page structure, updating local content requires some
state magic and passing a callback the bottom bar can invoke.

Closes #45