feat: add button to fetch original content
miniflux provides us 2 APIs: one to fetch the original content, and one
to update an entity. this first crude attempt downloads original content
from miniflux and, if that succeeds, then we PUT an update to that
entity and we replace the local content. due to entity page structure,
updating local content requires some state magic and passing a callback
the bottom bar can invoke